Configuration of spiking neurons, do not set `enable_digital`
`enable_digital` is already set in grenade the corresponding neurons
(compare
 https://gerrit.bioai.eu:9443/gitweb?p=grenade.git;a=blob;f=src/grenade/vx/execution_instance_config_visitor.cpp;h=3764837123a737050f0b1b29dee36154f5981bdc;hb=refs/heads/master;js=1#l274).
Since looping over the recordables is no longer needed, the check for
MADC recordings is also removed. This check is already performed when
defining a recording and an according test is present with
`test_record` in `pynn_brainscales2_test_population.py`.

Change-Id: Id72ccb781f1311cb393f0157c7091182f8babedd