Connections: introduce source and target selctors
This allows to connect neurons with multiple comaprtments. Currently we
only support the selection of the source/target via labels. In future
more complex selectors as proposed in the PyNN MC API could be
supported.

Since the selector argument is passed to `Connector` we have to
re-implement the `__init__` of all relevant connectors and add these
arguments there.

Change-Id: I31b79929d3c37c57489f33044418577a5ae94c7e