Use dict for madc_recording_sites
This is needed to later correctly assign the madc samples and times to
the corresponding recording sites. Previously a set() was used which
does not preserve order.

More detail: In `simulator.get_v()` we filter the samples for the
different recording sites and save them in the lists
`simulator.state.times` and `simulator.state.madc_samples`. When we
later extract the samples for a given population, we need to know in
which position the recorded samples are. I.e. we need to make sure that
we use the same order when we save the samples in `simulator.state` and
when we extract them.

Change-Id: I060f2e15451f8100f23e6b5d96731e7bea581c26